# batch reading for TBaskets